requirements.txt
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pillow==11.3.0
POT==0.9.6.post1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.16.2
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torchsummary==1.5.1
tqdm==4.67.1
typing_extensions==4.15.0